+++ src/main/java/ladder/domain/DrawResult.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 public class DrawResult {
     private final String result;
