+++ src/test/java/ladder/domain/DrawResultTest.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 import org.junit.jupiter.api.Test;
